packages/py/aitbc-crypto/pyproject.toml
[project]
name = "aitbc-crypto"
version = "0.1.0"
description = "AITBC cryptographic utilities"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.7.0",
    "pynacl>=1.5.0"
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
